A lot of work to optimize the rendering process. - Added VBO support; - added world map; - changed all doubles to floats; - added two worldmap-related options in SettingsWindow;
TODO: Rendering whole world from one VBO.

+++ source/db/AirportsDatabase.h
@@ -38,8 +38,8 @@ struct AirportRecord {
 	char	iata[8];
 	char	icao[8];
 	char	fir[8];
-	double	latitude;
-	double	longitude;
+	float	latitude;
+	float	longitude;
 	int	altitude;
 	
 };
